requirements.txt
# Existing
streamlit
prophet
numpy
pandas
yfinance
plotly
statsmodels

# New — Agentic Trading System
python-dotenv
websocket-client
alpaca-py
finnhub-python
fredapi
feedparser
groq
google-genai
aiohttp
requests
